+++ src/main/java/com/sandbox/company/service/SalaryEmployeeService.java
@@ -1,5 +1,6 @@
 package com.sandbox.company.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.sandbox.company.entity.SalaryEmployee;
@@ -12,7 +13,7 @@ public interface SalaryEmployeeService {
 
     SalaryEmployee saveSalaryEmployee(SalaryEmployee salaryEmployee, Long id);
 
-    SalaryEmployee updateSalaryEmployeeSalaryRate(double salaryRate, Long id);
+    SalaryEmployee updateSalaryEmployeeSalaryRate(BigDecimal salaryRate, Long id);
 
     SalaryEmployee updateSalaryEmployeeFullTime(boolean fullTime, Long id);
